+++ pkg/lib/web/login_link.go
@@ -1,4 +1,4 @@
-package webapp
+package web
 
 import "net/http"
